fix: IndustryNotFoundException 타입 불일치 문제 해결
- List<Long>을 List<Integer>로 변경하여 DTO 타입과 일치
- CorporationCreateDto의 industryIds 타입과 맞춤
- 컴파일 에러 해결

+++ src/main/java/com/newcodes7/small_town/corporation/exception/IndustryNotFoundException.java
@@ -8,7 +8,7 @@ public class IndustryNotFoundException extends CorporationException {
         super("INDUSTRY_NOT_FOUND", "업종을 찾을 수 없습니다. ID: %d", industryId);
     }
     
-    public IndustryNotFoundException(List<Long> industryIds) {
+    public IndustryNotFoundException(List<Integer> industryIds) {
         super("INDUSTRY_NOT_FOUND", "일부 업종을 찾을 수 없습니다. IDs: %s", industryIds.toString());
     }
 }
